@@ -5,7 +5,7 @@ homepage: https://github.com/GeekAbdelouahed/flutter-reaction-button
 repository: https://github.com/GeekAbdelouahed/flutter-reaction-button.git
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.6.0 <3.0.0"
 
 dependencies:
   flutter:
